ipts: Return altitude / azimuth in radians
The StylusData structure is not intended to end up as binary data,
so we can convert these values directly in the parser to prevent every
consumer needing their own conversion.

+++ src/ipts/data.hpp
@@ -21,8 +21,8 @@ struct StylusData {
 	u16 x = 0;
 	u16 y = 0;
 	u16 pressure = 0;
-	u16 altitude = 0;
-	u16 azimuth = 0;
+	f64 altitude = 0;
+	f64 azimuth = 0;
 	u32 serial = 0;
 };
